requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
distro==1.9.0
dnspython==2.8.0
groq==1.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.16.0
python-dotenv==1.2.2
python-telegram-bot==22.6
rich==14.3.3
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
